#651 Prevent, and fix existing illegal positions
Fixes the illegal position exploit. Resolves #651.
Prevents the illegal position from ever happening by checking the coordinates when a player teleports and canceling if the teleport would put them into an illegal position. Fixes existing illegal positions by checking coordinates on join, and sending the player to spawn if it is an illegal position.

I have tested the code and it seems to work fine.