Add Command Admin List
This Command would make it easier for ops to know our communitie's staff members, developers, Executive Seniors and server Owners and Co-Owners. It would also make it less spammy in the chat if a person starts to ask for our staff members, or admins not knowing the correct igns of our respected, high members.