Resolve linking issue
This loosely connects to the issue in question. However, because it revolves around the TFM release page, I figured it was appropriate to fix this issue also. The plus-symbol links to statistics rather than the actual link. Therefore, by removing the plus symbol you will be able to access the TFM releases page directly.