Make self verification system toggleable
This is great for if the password was indeed leaked senior admins may disable the self verification system until further notice. Please note that upon restarting the server it will need to disable again due to the config does not save. I may add a main config saving system soon.