Update Command_smite.java (Cosmetic Change)
This update makes smite messages stand out more in chat and to the smitten player (since even if a reason is specified, the smitten user wont know why they were smitten most of the time). It changed the default red text to a higher contrast/colour (yellow with the colour code §e/&e). This update makes it easier for people to view the reason for their smite and the smitten player theirself since the current colour (red) of smite reasons look dull and don't stand out.